requirements.txt
gunicorn
psycopg2
whitenoise
flask 
flask_restful
flask-jsonpify
flask-sqlalchemy
requests
beautifulsoup4
urllib
ssl
